# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WowoClean API Gateway",
        "description": "Dokumentasi API UAP TIS",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Endpoint Login",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@wowoclean.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil login & dapat token"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Ambil data user yang sedang login",
                "operationId": "e7fa93473c44547baf6183ba42322df2",
                "responses": {
                    "200": {
                        "description": "Data profile berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout & invalidate token",
                "operationId": "c4eddaf1a2e990cc90e4326c0d731d41",
                "responses": {
                    "200": {
                        "description": "Berhasil logout"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers": {
            "get": {
                "tags": [
                    "Containers V1"
                ],
                "summary": "Ambil semua data kontainer",
                "operationId": "c8a2aaa7f663413a114d18c5e173d11f",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data kontainer"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Containers V1"
                ],
                "summary": "Tambah kontainer baru (Admin only)",
                "operationId": "b3f02229a3d7b316446d1fa45d9c9604",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "container_id": {
                                        "type": "string",
                                        "example": "WH12345"
                                    },
                                    "waste_type": {
                                        "type": "string",
                                        "example": "Plastic"
                                    },
                                    "weight_kg": {
                                        "type": "number",
                                        "example": 200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data berhasil disimpan"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Hanya untuk Admin"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers/{id}/archive": {
            "patch": {
                "tags": [
                    "Containers V1"
                ],
                "summary": "Arsipkan kontainer (Admin only)",
                "operationId": "6bc059ac58eaca5134975156debacbaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status diubah menjadi Archived"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Hanya untuk Admin"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers/{id}": {
            "delete": {
                "tags": [
                    "Containers V1"
                ],
                "summary": "Hapus kontainer (Admin only)",
                "operationId": "526b15934fc7bf89ba909708ce954f44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Hanya untuk Admin"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers/search": {
            "get": {
                "tags": [
                    "Containers V1"
                ],
                "summary": "Cari kontainer berdasarkan filter",
                "operationId": "3d051fd37e25a62a7742c777cfb096c2",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Plastic"
                        }
                    },
                    {
                        "name": "min_weight",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil pencarian kontainer"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/gateway/containers/{id}/logs": {
            "get": {
                "tags": [
                    "Containers V1"
                ],
                "summary": "Ambil tracking logs kontainer",
                "operationId": "dc3d93f8c3bb74b22e2536236071786b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data logs berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Containers V1",
            "description": "Containers V1"
        }
    ]
}